add test runner
Realized I was using Python 2.7's unittest discover when I tried to use a Python 3 keyword for open() in one of my tests. Unfortunately, Python 3's unittest discovery module has a different implementation that flubs the import statements in my tests. testrunner.py is the solution, which coincidentally fixed an instance bug with the tkserializer unittest.