update to properly use modules
Previously I had trouble getting modules to work. This is because I
wasn't running it as a module, which you do in python by using the -m
flag, i.e.: python3 -m tenk.main. This wasn't written in the Python
documentation on modules, but I finally found the answer when I was on
google looking for a solution to an unrelated problem.

Other changes:
tenk/main.py: add documentation, clean up excessive comments
tests/stubs.py: remove as it is unnecessary