incorporate config files
I decided to use python's configparser and incorporate configuration
files. Currently there are two sections in the config file - PATHS and
NOTE CATEGORIES.

Now instead of using the user's home directory, the user can just choose
which directory tenk user files are stored in. The downside to this is
that the user must set up the paths before using the program.

I decided to incorporate note categories when I realized that I disliked
the original note categories I hard coded in (practiced, improved,
future) and if anyone were to use my program (haha) they probably
wouldn't like those categories either. So now categories are pulled from
the config file. If a user doesn't specify a category at the command
line, their input will be put in a node called notes rather than a
custom category node.

So the big plus in all of this is that my program has a clearer
separation of code and data than it did before.