use QSplitter rather than QHBoxLayout
Opted to use QSplitter in add_skill_to_main_win. It seemed like the more
appropriate tool for the job rather than adding a third layout
layer. Also remedied 'layout "" already has a parent' warning by moving
the addLayout line from add_skill_to_main_win. Qt was complaining it
because the line should only be called once but I was calling it
repeatedly since that function is called for each skill in a user's
skillset and when a user adds a new skill.