replace dictionary with OrderedDict for notes
It's better for notes to be stored in an ordered structure. It's quite
strange for the various note categories to show up in a different order
when viewing entries or entering notes for a new session via the GUI.

tenk/sessions.py: change __init__(), and deserialize() to reflect the
switch to OrderedDict to store note data.
tkcli/main.py: change note adding and viewing methods to reflect the
switch to OrderedDict.
tenk/users.py: change variable from 'time' to 'hours in add_time() for
better consistency.